Add discord-style italic and bold text formatting
Text surrounded by "**" as in **example** will be made bold, text
surrounded by '*' as in *example* will be made italic.

Unfortunately if bold-marked text is nested inside italic-marked text,
only the bold text will be made so. This is due to how chat messages are
parsed and "rebuilt" with formatting added and cannot be easily fixed.